requirements.txt
requests==2.21.0
beautifulsoup4==4.7.1
python_dateutil==2.8.0
